requirements.txt
PyYAML==6.0.1
requests==2.32.3
typer==0.12.3
art==6.2
inquirer==3.3.0